% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfume.R
\name{perfume}
\alias{perfume}
\title{This function calls \code{perfumeNames} defined by cpp}
\usage{
perfume()
}
\description{
This function calls \code{perfumeNames} defined by cpp
}
\details{
Perfume members introduce themselves on your console.
}

